tests/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
flake8==7.3.0
frozenlist==1.8.0
idna==3.11
iniconfig==2.3.0
mccabe==0.7.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
slack_bolt==1.26.0
slack_sdk==3.38.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
google-cloud-firestore==2.16.0